fix field_use having the wrong header
When I wrote the generator for the field_use images, I accidentally used
the obstacle's top_bg instead of the assist's top_bg. This came about
since I copied the code from the assist generator and forgot to change
it. When refactoring the part loading, I noticed I'd screwed this up,
and really this was a dumb mistake to make. Better part loading would
have caught this in a heartbeat, but because each generator had its own
parts storage, something like this was bound to happen sooner or later.